+++ source/draw.h
@@ -7,6 +7,6 @@ using namespace std;
 void drawBox(double x, double y, double width, double height, float a, float b, float c);
 void drawText(int x, int y, string text);
 void drawBox(double x, double y, double width, double height);
-void drawBox(double *pos);
+void drawBox(int *pos);
 void exitAll();
 #endif
